add option to override 'pixel is inside'-test
a new option `isInside` can be used to specify a function that is called instead of the default `inside`-function to determine if the given coordinate is considered inside the shape. This can be useful for images that only vary in their opacity but are otherwise completely black. 

As this is the only place where the source-image is actually accessed it can also be used to support ndarrays with shapes like `[width, height, 1]`.

Example:

```javascript

sourceImage = ndarray(imageData, [width, height, 4]);
var output = sdf(sourceImage, { isInside: function(array, x, y)  { return array.get(x,y,3) > 128; }}

```